# api/openapi.yaml
openapi: 3.0.3
info:
  title: Walwatch Auto-Renewal Keeper API
  description: API for managing blob auto-renewal on Walrus/Sui.
  version: 0.1.0
  contact:
    name: Walwatch
    url: https://walwatch.io
servers:
  - url: https://api.walwatch.io/api/v1
    description: Production
  - url: http://localhost:3001/api/v1
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
  parameters:
    OrgIdHeader:
      name: X-Org-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Organization UUID for scoping requests
...
  schemas:
...
    BlobRegistration:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
          nullable: true
        blobId:
          type: string
        name:
          type: string
          nullable: true
        sizeBytes:
          type: integer
          nullable: true
        contentType:
          type: string
          nullable: true
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        suiVaultId:
          type: string
          nullable: true
        ownerAddress:
          type: string
          nullable: true
        expiryEpoch:
          type: integer
          nullable: true
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateBlobInput:
      type: object
      required:
        - projectId
        - blobId
      properties:
        projectId:
          type: string
          format: uuid
        blobId:
          type: string
        name:
          type: string
        sizeBytes:
          type: integer
        contentType:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        suiVaultId:
          type: string
        ownerAddress:
          type: string
    BulkBlobInput:
      type: object
      required:
        - action
        - ids
      properties:
        action:
          type: string
          enum: [archive, activate, delete]
        ids:
          type: array
          items:
            type: string
            format: uuid
    BulkBlobResponse:
      type: object
      properties:
        message:
          type: string
        processed:
          type: integer
        skipped:
          type: integer

    # ── Policy ──
...
    PaginatedBlobResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BlobRegistration'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
paths:
...
  /projects/{projectId}/blobs:
    get:
      tags: [Projects]
      summary: List blobs in project
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated blobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBlobResponse'

  # ── Blobs ──
  /blobs:
    get:
      tags: [Blobs]
      summary: List blob registrations
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: project_id
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated blobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBlobResponse'
    post:
      tags: [Blobs]
      summary: Register blob
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlobInput'
      responses:
        '201':
          description: Blob registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlobRegistration'
  /blobs/bulk:
    post:
      tags: [Blobs]
      summary: Bulk blob action
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkBlobInput'
      responses:
        '200':
          description: Bulk action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkBlobResponse'
  /blobs/export:
    get:
      tags: [Blobs]
      summary: Export all blobs
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: All blob registrations
          content:
            application/json:
              schema:
                type: object
                properties:
                  blobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/BlobRegistration'
  /blobs/{blobId}:
    get:
      tags: [Blobs]
      summary: Get blob registration
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - name: blobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Blob registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlobRegistration'
    patch:
      tags: [Blobs]
      summary: Update blob registration
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - name: blobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Blob updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlobRegistration'
    delete:
      tags: [Blobs]
      summary: Delete blob registration
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - name: blobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Blob deleted

  # ── Policies ──
...